+++ PhysicsTools/Heppy/python/physicsobjects/Tau.py
@@ -1,5 +1,5 @@
 from PhysicsTools.Heppy.physicsobjects.Lepton import Lepton
-from PhysicsTools.Heppy.physicsobjects.TauDecayModes import tauDecayModes
+from PhysicsTools.Heppy.physicsutils.TauDecayModes import tauDecayModes
 import math
 
 cutsElectronMVA3Medium = [0.933,0.921,0.944,0.945,0.918,0.941,0.981,0.943,0.956,0.947,0.951,0.95,0.897,0.958,0.955,0.942]
